Always use alien indexing for projectile.
This should be fine for git projects, since the config
relies on git no matter what. On generic problems it will
most likely fail, since it relies on the Unix "find" command.
The external command can be customized however, so I will look
into that if it turns out to be a problem.