Fix terminal loading
This seems to actually be a bug in Emacs. The function
tty-run-terminal-initialization dynamically loads a lisp file
depending on the type of terminal, but it fails to strip all
extensions from the file name, which causes problems when the
extension happens to be ".el.gz", i.e. a gzipped file. We work around
this for now with a bit of advice hackery, but ideally this should be
fixed upstream.

Note that this problem is a direct consequence of the startup
optimization of temporarily setting file-name-handler-alist to nil
during startup, since tty-run-terminal-initialization runs before
emacs-startup-hook, which takes care of resetting the alist. So
another way to fix this would be to actually admit that this
optimization is probably more trouble than it is worth...